Debug Log 4/29/23
Created 2 new scripts, Grapple2 and Grapple3. These scripts are currently used to experiment with different ways to have a grappling process. Grapple3 is the one that is turned on in the level but due to upgrades may cause bugs down the line if not named properly. When this script is finished, Delete grapple  and grapple2 then rename grapple3 to grapple.
- Changed minimap camera settings to not cause bug errors
- Changed minimap camera settings to not be blue
- added anim boolean to player anim control that references the flying animation
- implemented flying animation into grapple script
-added a line renderer to the grapple script
- fixed grapple decrease amount

Issues: Grapple3 needs to be retested on controller and axis information for input may need to be changed.