Dash Temporary Fix 1
This changes the dash scirpt away from ray casting so its no longer maybe hit or not hit
This instead makes it so that if you collide with something, stop using rigid body add force by setting velocity to 0.
Finally, the model has a change in its collider- going from a box to a capsule now.
This allows the player to slowly get back to where they need to be when they go through the box.